Proof of concept: functional timer
At this stage the plugin is able to print some text from a bash script.
As a proof of concept the text printed is a simple 5 sec loop.
Since the code structure has been copied from tmux-spotify-plugin, some
code which is probably unecessary to us is still in place to have a
working plugin (mainly helpers.sh).